# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ynov API for YnovImmo"
  version: "1.0.0"
  title: "Ynov API pour l'application YnovImmo"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /real-estate-image:
    get:
      tags:
      - "real-estate-image"
      summary: "get all real-estate-image by pagination"
      description: ""
      operationId: "real_estate_imageGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateImagePagination"
      x-swagger-router-controller: "RealEstateImage"
    post:
      tags:
      - "real-estate-image"
      summary: "create real-estate-image"
      description: "create real-estate-image"
      operationId: "real_estate_imagePOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-image"
        required: true
        schema:
          $ref: "#/definitions/RealEstateImage"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateImage"
    patch:
      tags:
      - "real-estate-image"
      summary: "update real-estate-image"
      description: "update real-estate-image"
      operationId: "real_estate_imagePATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-image"
        required: true
        schema:
          $ref: "#/definitions/RealEstateImage"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateImage"
  /real-estate-image/{Id}:
    get:
      tags:
      - "real-estate-image"
      summary: "get a real-estate-image by ID"
      description: "get a real-estate-image by ID"
      operationId: "real_estate_imageIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-image to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateImage"
      x-swagger-router-controller: "RealEstateImage"
    delete:
      tags:
      - "real-estate-image"
      summary: "Destroy a real-estate-image by ID"
      description: "delete a real-estate-image by ID"
      operationId: "real_estate_imageIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-image to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateImage"
  /real-estate-propal:
    get:
      tags:
      - "real-estate-propal"
      summary: "get all real-estate-propal by pagination"
      description: ""
      operationId: "real_estate_propalGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstatePropalPagination"
      x-swagger-router-controller: "RealEstatePropal"
    post:
      tags:
      - "real-estate-propal"
      summary: "create real-estate-propal"
      description: "create real-estate-propal"
      operationId: "real_estate_propalPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propal"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropal"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstatePropal"
    patch:
      tags:
      - "real-estate-propal"
      summary: "update real-estate-propal"
      description: "update real-estate-propal"
      operationId: "real_estate_propalPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propal"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropal"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstatePropal"
  /real-estate-propal/{Id}:
    get:
      tags:
      - "real-estate-propal"
      summary: "get a real-estate-propal by ID"
      description: "get a real-estate-propal by ID"
      operationId: "real_estate_propalIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propal to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstatePropal"
      x-swagger-router-controller: "RealEstatePropal"
    delete:
      tags:
      - "real-estate-propal"
      summary: "Destroy a real-estate-propal by ID"
      description: "delete a real-estate-propal by ID"
      operationId: "real_estate_propalIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propal to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstatePropal"
  /real-estate-propals-buyer-indication:
    get:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "get all real-estate-propals-buyer-indication by pagination"
      description: ""
      operationId: "real_estate_propals_buyer_indicationGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstatePropalsBuyerIndicationPagination"
      x-swagger-router-controller: "RealEstatePropalsBuyerIndication"
    post:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "create real-estate-propals-buyer-indication"
      description: "create real-estate-propals-buyer-indication"
      operationId: "real_estate_propals_buyer_indicationPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propals-buyer-indication"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropalsBuyerIndication"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstatePropalsBuyerIndication"
    patch:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "update real-estate-propals-buyer-indication"
      description: "update real-estate-propals-buyer-indication"
      operationId: "real_estate_propals_buyer_indicationPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-propals-buyer-indication"
        required: true
        schema:
          $ref: "#/definitions/RealEstatePropalsBuyerIndication"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstatePropalsBuyerIndication"
  /real-estate-propals-buyer-indication/{Id}:
    get:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "get a real-estate-propals-buyer-indication by ID"
      description: "get a real-estate-propals-buyer-indication by ID"
      operationId: "real_estate_propals_buyer_indicationIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propals-buyer-indication to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstatePropalsBuyerIndication"
      x-swagger-router-controller: "RealEstatePropalsBuyerIndication"
    delete:
      tags:
      - "real-estate-propals-buyer-indication"
      summary: "Destroy a real-estate-propals-buyer-indication by ID"
      description: "delete a real-estate-propals-buyer-indication by ID"
      operationId: "real_estate_propals_buyer_indicationIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-propals-buyer-indication to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstatePropalsBuyerIndication"
...
definitions:
...
  RealEstateImage:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de l'image"
      id_real_estate:
        type: "integer"
        format: "int64"
        description: "Id du bien"
      url:
        type: "string"
        format: "string"
        description: "Url de l'image"
    example:
      id: 5
      id_real_estate: 2
      url: "url"
  RealEstateImagePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateImage"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - id: 5
        id_real_estate: 2
        url: "url"
      - id: 5
        id_real_estate: 2
        url: "url"
  RealEstatePropal:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de la proposition"
      id_booker:
        type: "integer"
        format: "int64"
        description: "Id du booker"
      id_buyer:
        type: "integer"
        format: "int64"
        description: "Id du visitor"
      id_real_estate:
        type: "integer"
        format: "int64"
        description: "Id du bien"
      id_vendor:
        type: "integer"
        format: "int64"
        description: "Id du propriétaire"
      price:
        type: "number"
        format: "float"
        description: "Montant de la proposition"
      buyer_first_name:
        type: "string"
        format: "string"
        description: "Prénom de l'acheteur"
      buyer_last_name:
        type: "string"
        format: "string"
        description: "Nom de l'acheteur"
      buyer_address:
        type: "string"
        format: "string"
        description: "Adresse de l'acheteur"
      buyer_zip_code:
        type: "string"
        format: "string"
        description: "Code postal de l'acheteur"
      buyer_city:
        type: "string"
        format: "string"
        description: "Ville de l'acheteur"
      expiration_date:
        type: "string"
        format: "date-time"
        description: "Date d'expiration de l'offre"
      is_accepted:
        type: "integer"
        format: "int64"
        description: "Est accepté?"
    example:
      is_accepted: 4
      buyer_first_name: "buyer_first_name"
      buyer_address: "buyer_address"
      id_real_estate: 9
      expiration_date: "2000-01-23T04:56:07.000+00:00"
      id_buyer: 7
      buyer_zip_code: "buyer_zip_code"
      buyer_city: "buyer_city"
      price: 2.027123
      id_booker: 2
      id_vendor: 3
      id: 5
      buyer_last_name: "buyer_last_name"
  RealEstatePropalPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstatePropal"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - is_accepted: 4
        buyer_first_name: "buyer_first_name"
        buyer_address: "buyer_address"
        id_real_estate: 9
        expiration_date: "2000-01-23T04:56:07.000+00:00"
        id_buyer: 7
        buyer_zip_code: "buyer_zip_code"
        buyer_city: "buyer_city"
        price: 2.027123
        id_booker: 2
        id_vendor: 3
        id: 5
        buyer_last_name: "buyer_last_name"
      - is_accepted: 4
        buyer_first_name: "buyer_first_name"
        buyer_address: "buyer_address"
        id_real_estate: 9
        expiration_date: "2000-01-23T04:56:07.000+00:00"
        id_buyer: 7
        buyer_zip_code: "buyer_zip_code"
        buyer_city: "buyer_city"
        price: 2.027123
        id_booker: 2
        id_vendor: 3
        id: 5
        buyer_last_name: "buyer_last_name"
  RealEstatePropalsBuyerIndication:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de l'indication"
      id_propal:
        type: "integer"
        format: "int64"
        description: "Id de la proposition"
      id_buyer:
        type: "integer"
        format: "int64"
        description: "Id de l'acheteur"
      id_vendor:
        type: "integer"
        format: "int64"
        description: "Id du vendeur"
      has_total_in_cash:
        type: "integer"
        format: "int64"
        description: "a le total en cash ?"
      has_financial_support:
        type: "integer"
        format: "int64"
        description: "a un apport financier ?"
      has_big_financial_support:
        type: "integer"
        format: "int64"
        description: "a un apport financier important ? (>30% du bien)"
      has_very_big_financial_support:
        type: "integer"
        format: "int64"
        description: "a un apport financier très important ? (> 70% du bien)"
      want_a_long_sell:
        type: "integer"
        format: "int64"
        description: "veut une vente longue ? (+ de 6 mois)"
    example:
      id_propal: 2
      has_total_in_cash: 3
      has_big_financial_support: 4
      id_vendor: 9
      id: 5
      has_financial_support: 2
      want_a_long_sell: 1
      id_buyer: 7
      has_very_big_financial_support: 7
  RealEstatePropalsBuyerIndicationPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstatePropalsBuyerIndication"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - id_propal: 2
        has_total_in_cash: 3
        has_big_financial_support: 4
        id_vendor: 9
        id: 5
        has_financial_support: 2
        want_a_long_sell: 1
        id_buyer: 7
        has_very_big_financial_support: 7
      - id_propal: 2
        has_total_in_cash: 3
        has_big_financial_support: 4
        id_vendor: 9
        id: 5
        has_financial_support: 2
        want_a_long_sell: 1
        id_buyer: 7
        has_very_big_financial_support: 7
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
    example:
      msg: "msg"
      code: 0
...